+++ package.json
@@ -13,7 +13,8 @@
     "@splidejs/splide": "^4.1.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.8.0"
+    "react-icons": "^4.8.0",
+    "react-router-dom": "^6.11.0"
   },
   "devDependencies": {
     "@babel/core": "^7.21.4",
