--- src/Components/About/index.jsx
@@ -1,5 +1,0 @@
-const index = () => {
-  return <div>About</div>;
-};
-
-export default index;
