+++ src/Components/Home/index.jsx
@@ -1,0 +1,13 @@
+import Landing from "./Landing.jsx";
+import Schedule from "./Schedule.jsx";
+
+const Home = () => {
+    return (
+        <>
+            <Landing/>
+            <Schedule />
+        </>
+    );
+};
+
+export default Home;
