+++ src/Components/News/index.jsx
@@ -1,0 +1,5 @@
+const News = () => {
+  return <div>News</div>;
+};
+
+export default News;
