+++ src/Components/OurClub/index.jsx
@@ -1,0 +1,5 @@
+const index = () => {
+  return <div>About</div>;
+};
+
+export default index;
