+++ src/Components/Team/index.jsx
@@ -1,0 +1,5 @@
+const Team = () => {
+  return <div>About</div>;
+};
+
+export default Team;
